requirements.txt
flask 
transformers 
torch 
paho-mqtt 
pandas 
numpy 
spacy 
requests 
